Change routing design. Fix according CR.
