+++ tutor/controllers/hello.py
@@ -1,4 +1,6 @@
 from flask import render_template
+from .. import app
+
  
 def hello():
     return render_template('index.html')
