+++ tutor/controllers/model.py
@@ -3,15 +3,14 @@ from .. import db
 from flask import jsonify
 
 
-class ModelController:
-    def getAllModels():
+def getAllModels():
         models = Model.query.all()
         return jsonify([e.serialize() for e in models])
 
-    def createNewModel(modelName):
-        model = Model(name=modelName)
-        db.session.add(model)
-        db.session.commit()
-        return jsonify({'id' : model.id, 'status' : 200})
 
-
+def createNewModel(modelName):
+    model = Model(name=modelName)
+    db.session.add(model)
+    db.session.commit()
+    return jsonify({'id' : model.id, 'status' : 200})
+
