+++ tutor/routes/hello.py
@@ -1,0 +1,7 @@
+from .. import app
+from ..controllers import hello
+
+
+@app.route('/')
+def hello_route():
+    return hello.hello()
